Cleanup
Closes: #16475
