+++ docs/antora/antora-assembler.yml
@@ -2,5 +2,8 @@ root_level: 0
 component_versions: '*'
 section_merge_strategy: fuse
 build:
+  # only needed for original pdf-extension
+  keep_aggregate_source: false
+  # only needed for original pdf-extension
   command: 'true'
-  keep_aggregate_source: true
+  publish: false
