+++ docs/antora/ccutil/guides-server/master.adoc
@@ -3,114 +3,5 @@
 :numbered:
 :linkattrs:
 
-// TODO: generate attributes below from antora-playbook.yml, or even better capture them when Antora Assembler passes tem on the command line.
-
-:project_name: Keycloak
-:project_name_full: Keycloak
-:project_community: true
-:project_product: false
-:project_images: keycloak-images
-:project_versionLinks: latest
-:project_doc_base_url: 'https://www.keycloak.org/docs/{project_versionLinks}'
-:project_doc_base_url_latest: https://www.keycloak.org/docs/latest
-:project_dirref: KEYCLOAK_HOME
-:project_openshift_product_name: Keycloak for OpenShift
-:project_version: 20.0.0
-:project_versionMvn: 20.0.0
-:project_versionNpm: 20.0.0
-:project_versionDoc: 20.0
-:kc_realms_path: /realms
-:kc_admins_path: /admin
-:kc_js_path: /js
-:kc_base_path: ''
-:project_operator: Keycloak Operator
-:operatorRepo_link: https://github.com/keycloak/keycloak-operator
-:application_monitoring_operator: Application Monitoring Operator
-:application_monitoring_operator_installation_link: https://github.com/integr8ly/application-monitoring-operator#installation
-:create_cmd: kubectl apply
-:create_cmd_brief: kubectl
-:jdgserver_name: Infinispan
-:jdgserver_version: 9.4.19
-:jdgserver_version_latest: 11.0.9
-:jdgserver_crossdcdocs_link: https://infinispan.org/docs/11.0.x/titles/xsite/xsite.html
-:subsystem_undertow_xml_urn: urn:jboss:domain:undertow:12.0
-:subsystem_infinispan_xml_urn: urn:jboss:domain:infinispan:12.0
-:subsystem_datasources_xml_urn: urn:jboss:domain:datasources:6.0
-:saml_adapter_xsd_urn: https://www.keycloak.org/schema/keycloak_saml_adapter_1_10.xsd
-:tech_feature_disabled: false
-:quickstartRepo_link: https://github.com/keycloak/keycloak-quickstarts
-:quickstartRepo_name: Keycloak Quickstarts Repository
-:quickstartRepo_dir: keycloak-quickstarts
-:authorizationguide_name: Authorization Services Guide
-:authorizationguide_name_short: Authorization Services
-:authorizationguide_link: '{project_doc_base_url}/authorization_services/'
-:authorizationguide_link_latest: '{project_doc_base_url_latest}/authorization_services/'
-:adapterguide_name: Securing Applications and Services Guide
-:adapterguide_name_short: Securing Apps
-:adapterguide_link: '{project_doc_base_url}/securing_apps/'
-:adapterguide_link_js_adapter: '{adapterguide_link}#_javascript_adapter'
-:adapterguide_link_latest: '{project_doc_base_url_latest}/securing_apps/'
-:adapterguide_logout_link: '{adapterguide_link}#_java_adapter_logout'
-:adminguide_name: Server Administration Guide
-:adminguide_name_short: Server Administration
-:adminguide_link: '{project_doc_base_url}/server_admin/'
-:adminguide_link_latest: '{project_doc_base_url_latest}/server_admin/'
-:adminguide_bruteforce_name: 'Password guess: brute force attacks'
-:adminguide_bruteforce_link: '{adminguide_link}#password-guess-brute-force-attacks'
-:adminguide_timeouts_name: Timeouts
-:adminguide_timeouts_link: '{adminguide_link}#_timeouts'
-:adminguide_clearcache_name: Clearing Server Caches
-:adminguide_clearcache_link: '{adminguide_link}#_clear-cache'
-:apidocs_name: API Documentation
-:apidocs_link: https://www.keycloak.org/docs/{project_versionDoc}/api_documentation/
-:developerguide_name: Server Developer Guide
-:developerguide_name_short: Server Developer
-:developerguide_link: '{project_doc_base_url}/server_development/'
-:developerguide_link_latest: '{project_doc_base_url_latest}/server_development/'
-:developerguide_deploying_themes: '{developerguide_link}#deploying-themes'
-:developerguide_actiontoken_name: Action Token SPI
-:developerguide_actiontoken_link: '{developerguide_link}#_action_token_spi'
-:developerguide_jsproviders_name: JavaScript Providers
-:developerguide_jsproviders_link: '{developerguide_link}#_script_providers'
-:gettingstarted_name: Getting Started Guide
-:gettingstarted_name_short: Getting Started
-:gettingstarted_link: https://www.keycloak.org/guides#getting-started
-:gettingstarted_link_latest: https://www.keycloak.org/guides#getting-started
-:upgradingguide_name: Upgrading Guide
-:upgradingguide_name_short: Upgrading
-:upgradingguide_link: '{project_doc_base_url}/upgrading/'
-:upgradingguide_link_latest: '{project_doc_base_url_latest}/upgrading/'
-:releasenotes_name: Release Notes
-:releasenotes_name_short: '{releasenotes_name}'
-:releasenotes_link: '{project_doc_base_url}/release_notes/'
-:releasenotes_link_latest: '{project_doc_base_url_latest}/release_notes/'
-:apidocs_javadocs_name: JavaDocs Documentation
-:apidocs_javadocs_link: https://www.keycloak.org/docs-api/{project_versionDoc}/javadocs/
-:apidocs_adminrest_name: Administration REST API
-:apidocs_adminrest_link: https://www.keycloak.org/docs-api/{project_versionDoc}/rest-api/
-:appserver_name: WildFly
-:appserver_dirref: WILDFLY_HOME
-:appserver_version: 23
-:appserver_doc_base_url: http://docs.wildfly.org/{appserver_version}
-:appserver_socket_link: '{appserver_doc_base_url}/Admin_Guide.html#Interfaces_and_ports'
-:appserver_socket_name: '{appserver_name} {appserver_version} Documentation'
-:appserver_jgroups_link: '{appserver_doc_base_url}/High_Availability_Guide.html#JGroups_Subsystem'
-:appserver_jgroups_name: '{appserver_name} {appserver_version} Documentation'
-:appserver_jpa_link: '{appserver_doc_base_url}/Developer_Guide.html#hibernate-properties'
-:appserver_jpa_name: '{appserver_name} {appserver_version} Development Guide'
-:appserver_network_link: '{appserver_doc_base_url}/Admin_Guide.html#Interfaces_and_ports'
-:appserver_network_name: '{appserver_name} {appserver_version} Documentation'
-:appserver_datasource_link: '{appserver_doc_base_url}/Admin_Guide.html#DataSource'
-:appserver_datasource_name: '{appserver_name} {appserver_version} Documentation'
-:appserver_caching_link: '{appserver_doc_base_url}/High_Availability_Guide.html#Infinispan_Subsystem'
-:appserver_caching_name: '{appserver_name} {appserver_version} Documentation'
-:appserver_admindoc_link: '{appserver_doc_base_url}/Admin_Guide.html'
-:appserver_admindoc_name: '{appserver_name} {appserver_version} Documentation'
-:appserver_loadbalancer_link: '{appserver_doc_base_url}/High_Availability_Guide.html'
-:appserver_loadbalancer_name: '{appserver_name} {appserver_version} Documentation'
-:standalone: ''
-:api-management: ~
-:on-prem: ''
-
-include::20.0/guides-keycloak-server.adoc[]
+include::20.0/guides-keycloak-server-ccutil.adoc[]
